Move corestore.js to test.
